move LoopTree to analyse/loop_analysis.py
