+++ loki/transformations/utilities.py
@@ -10,7 +10,7 @@ Collection of utility routines to deal with general language conversion.
 """
 
 import platform
-from collections import defaultdict, deque
+from collections import defaultdict
 from pymbolic.primitives import Expression
 from loki.expression import (
     symbols as sym, SubstituteExpressionsMapper, ExpressionRetriever,
@@ -20,7 +20,7 @@ from loki.ir import (
     nodes as ir, Import, TypeDef, VariableDeclaration, is_loki_pragma,
     StatementFunction, Transformer, FindNodes, FindVariables,
     FindInlineCalls, FindLiterals, SubstituteExpressions,
-    ExpressionFinder, Visitor
+    ExpressionFinder
 )
 from loki.module import Module
 from loki.subroutine import Subroutine
@@ -28,6 +28,7 @@ from loki.tools import CaseInsensitiveDict, as_tuple
 from loki.types import SymbolAttributes, BasicType, DerivedType, ProcedureType
 from loki.config import config_override
 from loki.logging import warning
+from loki.analyse import get_loop_tree
 
 
 __all__ = [
@@ -586,117 +587,6 @@ def get_loop_bounds(routine, dimension):
     return bounds
 
 
-class LoopTree:
-    """
-    A data structure that stores nested loops in a forest of trees.
-
-    Parameters
-    ----------
-    roots : list of :any:`LoopTree.TreeNode`
-        top level loops in the ir section of the LoopTree
-    loop_map : dict of (:any:`ir.Loop`, :any:`TreeNode`)
-        dictionary with key-value pairs of :any:`ir.Loop` nodes in the tree
-        and corresponding :any:`TreeNode` nodes
-    """
-
-    class TreeNode:
-        """
-        Internal node class for nodes in the LoopTree
-
-        Parameters
-        ----------
-        loop : :any:`ir.Loop`
-            :any:`ir.Loop` of the tree node
-        parent : :any:`LoopTree.Treenode` or None
-            parent node in the LoopTree
-        parent : list of :any:`LoopTree.Treenode`
-            child nodes in the LoopTree
-        depth : int
-            nesting depth of the node (0 if root)
-        """
-
-        def __init__(self, loop: ir.Loop, parent=None):
-            self.loop = loop
-            self.parent = parent
-            self.children = []
-            self.depth = 0 if parent is None else parent.depth+1
-
-            if parent:
-                parent.children.append(self)
-
-    def __init__(self):
-        self.roots = []
-        self.loop_map = {}
-
-    def add_node(self, loop: ir.Loop, parent: TreeNode = None):
-        """
-        Helper function to add a loop to the LoopTree
-        """
-        tree_node = self.TreeNode(loop, parent)
-        if parent is None:
-            self.roots.append(tree_node)
-        self.loop_map[loop] = tree_node
-        return tree_node
-
-    def get_tree_node(self, loop: ir.Loop):
-        """
-        Get LoopTree node corresponding to Loki IR Loop node.
-        """
-        return self.loop_map.get(loop)
-
-    def walk_depth_first(self, pre_order=True):
-        """
-        Generator for depth first traversal of the loop tree.
-
-        Parameters
-        ----------
-        pre_order : `bool`
-            Order of depth first traversal, if `True` parent nodes are visited before children.
-            default: True
-        """
-        def visit(tree_node):
-            if pre_order:
-                yield tree_node
-
-            for child in tree_node.children:
-                yield from visit(child)
-
-            if not pre_order:
-                yield tree_node
-
-        for root in self.roots:
-            yield from visit(root)
-
-    def walk_breadth_first(self):
-        """
-        Generator for breadth first traversal of the loop tree.
-        """
-        queue = deque(self.roots)
-        while queue:
-            tree_node = queue.popleft()
-            yield tree_node
-            queue.extend(tree_node.children)
-
-
-def get_loop_tree(region: ir.Node):
-    """
-    Function that construct a loop tree with all loops in an IR region
-    """
-    class LoopTreeBuilder(Visitor):
-        def __init__(self):
-            super().__init__()
-            self.loop_tree = LoopTree()
-
-        def visit_Loop(self, loop, **kwargs):
-            parent = kwargs.pop('parent', None)
-            tree_node = self.loop_tree.add_node(loop, parent)
-
-            for o in loop.children:
-                self.visit(o, parent=tree_node, **kwargs)
-
-    tree_builder = LoopTreeBuilder()
-    tree_builder.visit(region)
-    return tree_builder.loop_tree
 
 
 def is_pragma_driver_loop(loop):
@@ -833,7 +723,7 @@ def find_driver_loops(section, targets):
             if is_target_driver:
                 if nested_pragma_driver:
                     warning("[Loki::find_driver_loops] Nested pragma marked driver loop inside loop"
-                            f"with target call (skipping {node.loop}")
+                            f" with target call (skipping {node.loop}")
                     return True
 
                 target_driver_loops.append(node)
